+++ src/main/java/racingcar/strategy/RacingNumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar.utils;
+package racingcar.strategy;
 
 public interface RacingNumberGenerator {
